Created file parse_quotes to have things related to quoted strings working independently from echo
