+++ Makefile
@@ -43,6 +43,7 @@ MINISHELL_SRC= 							\
 	src/commands/echo_command.c			\
 	src/commands/echo_utils.c			\
 	src/commands/quotes.c				\
+	src/commands/parse_quotes.c			\
 	src/commands/buffer.c				\
 	src/commands/pwd_command.c			\
 	src/commands/environment_commands.c	\
