+++ src/commands/quotes.h
@@ -1,6 +1,8 @@
 #ifndef QUOTES_H
 # define QUOTES_H
 # include <defines.h>
+# include <commands/commands.h>
+# include <commands/buffer.h>
 
 typedef struct s_quotes_position
 {
@@ -20,5 +22,6 @@ t_quotes_position	get_quotes_positions(const char *input);
 t_bool				is_double_quote(char c);
 t_bool				is_single_quote(char c);
 t_bool				is_quote(char c);
+t_arg				parse_str_with_quotes(t_arg arg, t_buffer *buffer);
 
 #endif
